requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
channels==4.1.0
channels-redis==4.2.0
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
msgpack==1.0.8
PyJWT==2.8.0
redis==5.0.6
sqlparse==0.5.0
typing-extensions==4.12.2
